+++ CHANGELOG.md
@@ -18,7 +18,8 @@ Bugs:
 
 Features:
 
-- None
+- `[utils.shell]` Support `suppress` meta option on `Prompt` to suppress user input.
+    - [Issue #621](https://github.com/datafolklabs/cement/issues/621)
 
 
 Refactoring: